sync: microsoft/calculator v2606 — C++ 引擎 + C# FFI + ViewModel 三层同步
C++ 引擎层:
- 新增 NumberFormattingUtils.h/cpp, sal_cross_platform.h, winerror_cross_platform.h
- CalcEngine.h: enum→enum class NUM_WIDTH, IsInputEmpty(), groupDigitsPerRadix
- ICalcDisplay.h: InputChanged() 回调
- CalculatorManager.h/cpp: InputChanged, IsInputEmpty, GetDisplayCommandsSnapshot
- calc.cpp: s_engineStrings 优化, LoadEngineStrings 简化
- 修复 6 处编译错误 (CalcInput.IsEmpty, NUM_WIDTH 索引, 函数签名等)
- build_linux.sh: c++1z → c++20

C# FFI 层:
- CalculatorManager.cs: ICalcDisplay.InputChanged, IsInputEmpty, GetDisplayCommandsSnapshot
- CalculatorManager.Interop.cs: P/Invoke 声明, InputChangedCallbackFunc delegate
- CCalcManager.h/.cpp: InputChangedFunc typedef, 导出函数实现
- CalcManager.js: WASM InputChanged 回调注册

ViewModel 层:
- StandardCalculatorViewModel.OnInputChanged() 实现 guard clause
- 空输入时 DisplayValue → "0"，非空时作为纯通知回调

验证: C++ x64✅ arm64✅, .NET 0错误0警告, 应用正常启动

+++ src/CalcManager/CEngine/scidisp.cpp
@@ -66,10 +66,10 @@ CalcEngine::Rational CCalcEngine::TruncateNumForIntMath(CalcEngine::Rational con
     {
         // if negative make positive by doing a twos complement
         result = -(result)-1;
-        result ^= m_chopNumbers[m_numwidth];
+        result ^= m_chopNumbers[static_cast<size_t>(m_numwidth)];
     }
 
-    result &= m_chopNumbers[m_numwidth];
+    result &= m_chopNumbers[static_cast<size_t>(m_numwidth)];
 
     return result;
 }
